Cleanup
Some cleanup to make things a bit easier to understand.
