+++ ClassicMode.xaml
@@ -2,10 +2,10 @@
     x:Class="Files.ClassicMode"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local1="using:ItemListPresenter"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:UILib="using:Microsoft.UI.Xaml.Controls"
+    xmlns:local1="using:Files.Filesystem"
     xmlns:Windows10version1809="http://schemas.microsoft.com/winfx/2006/xaml/presentation?IsApiContractPresent(Windows.Foundation.UniversalApiContract, 7)"
     mc:Ignorable="d"
     Name="ClassicModePage"
