+++ MainPage.xaml.cs
@@ -1,5 +1,4 @@
-﻿using ItemListPresenter;
-using System;
+﻿using System;
 using System.Collections;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
@@ -15,6 +14,7 @@ using Windows.UI.ViewManagement;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Media.Animation;
+using Files.Filesystem;
 
 namespace Files
 {
@@ -94,7 +94,7 @@ namespace Files
                     {
                         // Debug.WriteLine("{0} not found", let);
                     }
-                    
+                
                 }
             }
         }
