+++ Navigation/ForwardState.cs
@@ -1,0 +1,34 @@
+﻿using System.ComponentModel;
+
+namespace Files.Navigation
+{
+    public class ForwardState : INotifyPropertyChanged
+    {
+
+
+        public bool _isEnabled;
+        public bool isEnabled
+        {
+            get
+            {
+                return _isEnabled;
+            }
+
+            set
+            {
+                if (value != _isEnabled)
+                {
+                    _isEnabled = value;
+                    NotifyPropertyChanged("isEnabled");
+                }
+            }
+        }
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        private void NotifyPropertyChanged(string info)
+        {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(info));
+        }
+
+    }
+}
